pyproject.toml
[project]
name = "trading-crew"
version = "0.1.0"
description = "Multi-agent crypto trading system powered by CrewAI"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
authors = [{ name = "trading-crew contributors" }]
keywords = ["trading", "crypto", "crewai", "multi-agent", "ccxt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "crewai>=1.0,<2.0",
    "ccxt>=4.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pandas>=2.0",
    "numpy>=1.26",
    "pandas-ta>=0.3",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.8",
    "mypy>=1.11",
    "pre-commit>=3.8",
    "detect-secrets>=1.5",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
]
notifications = [
    "python-telegram-bot>=21.0",
]

[project.scripts]
trading-crew = "trading_crew.main:main"

[project.urls]
Homepage = "https://github.com/your-org/trading-crew"
Documentation = "https://your-org.github.io/trading-crew"
Repository = "https://github.com/your-org/trading-crew"
Issues = "https://github.com/your-org/trading-crew/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trading_crew"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["trading_crew"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests (fast, no external deps)",
    "integration: Integration tests (may need DB or network)",
    "backtest: Backtesting tests (may be slow)",
]
